Application/server/requirements.txt
flask
flask_cors
rembg
opencv-python
numpy
pillow